Add files via upload
New version of python forensic script. Should look much cleaner.
Still haven't figure out how to save the output from python. If you launch the application as such:
    python3 fsv2.py |& tee output.txt 
it should let you put the path in as normal and then it'll save the output to a .txt file for your viewing. You can call the file whatever you like. Also it might take a second or to for the script to get going after you enter the path :)